Pushed LWC, Apex, Apex Trigger
Lwc that creates a carousel of every product thats active and has DisplayUrl with the use of an Apex Class and SOQL query.
Apex class that adds a standard price of 100 everytime product is created with the use of an apex trigger.

Declarative customization: new page layout field for product : DisplayUrl field.